+++ frontend/src/components/admin/AuditLogsCard.vue
